Move custom exceptions to util module
