+++ lib/taurus/qt/qtgui/taurusgui/config_loader/jsonconf.py
@@ -27,7 +27,9 @@ import json
 import os
 
 from taurus.qt.qtgui.taurusgui.config_loader.abstract import (
-    AbstractConfigLoader, ConfigLoaderError)
+    AbstractConfigLoader
+)
+from taurus.qt.qtgui.taurusgui.config_loader.util import ConfigLoaderError
 from taurus.qt.qtgui.taurusgui.utils import (AppletDescription, ExternalApp,
                                              PanelDescription,
                                              ToolBarDescription)
