+++ lib/taurus/qt/qtgui/taurusgui/config_loader/util.py
@@ -34,7 +34,27 @@ import pkg_resources
 from taurus import warning
 
 
-__all__ = ["getLoaders"]
+__all__ = ["getLoaders", "ConfigLoaderError", "HookLoaderError"]
+
+
+class ConfigLoaderError(Exception):
+    """
+    Base exception raised by ConfigLoader
+    """
+
+    def __init__(self, message):
+        message = "Exception raised while loading configuration: " + message
+        super(ConfigLoaderError, self).__init__(message)
+
+
+class HookLoaderError(ConfigLoaderError):
+    """
+    Exception raised by config loader hooks
+    """
+
+    def __init__(self, message):
+        message = "Exception raised while executing config loader hookhook: " + message
+        super(HookLoaderError, self).__init__(message)
 
 
 def getLoaders(confname):
