+++ lib/taurus/qt/qtgui/taurusgui/config_loader/xmlconf.py
@@ -27,7 +27,9 @@ import os
 
 from lxml import etree
 from taurus.qt.qtgui.taurusgui.config_loader.abstract import (
-    AbstractConfigLoader, ConfigLoaderError)
+    AbstractConfigLoader
+)
+from taurus.qt.qtgui.taurusgui.config_loader.util import ConfigLoaderError
 from taurus.qt.qtgui.taurusgui.utils import (AppletDescription, ExternalApp,
                                              PanelDescription,
                                              ToolBarDescription)
